Add note about single section pages to the README
Section pages (other than projects.md and about.md which have custom
layouts) must be named `_index.html` in order to declare a branch bundle
and therefore a SiteSection. The default layout for these pages is
`list.html` and needs to be forced to `single.html` in order to display
as a single page.